Changes "Exercise Progress" table in user profile to use our api.
This is a proof of concept, and paves the way for us to do this with more and more of the app.

I've tested the best I can and the behavior seems identical, and I think this is a better approach in the long run. But in the short run, this makes loading that page slightly slower, since the response payload goes from 94k to 277k.