Initial handlebars prototype.
Only works in dev-mode - serves the templates as inline <script> tags.
One caveat: doing so requires the server to read the contents of the static files that contains the template source. However, GAE does not allow reading from files that are marked as static in app.yaml, since those files will ultimately get served off a static file server, and not part of the main package. A symlink 'clienttemplates' is made to get around this, but it should be noted that that solution will NOT work for production.

An upcoming CL will be made to compile the templates for production, and serve it along with the JS package it belongs to (this is nicer since it since it will have better caching properties anyways)