Change the way user exercice cache/graph building is done.
User exercise information has gotten large enough that fetching a reasonably sized list of students' exercise information causes the GAE instance to exceed memory thresholds.

This changes that logic so that user exercise information is fetched in batches (of 5). This will make things slightly slower when we change the schema of the user exercise cache objects, but prevent it from erroring out.