Introduce a testbase.js
This allows us to create standalone HTML files to aid with and encourage the development of re-usable JavaScript components.

Components can be built independent of the server changes and demoed (if it's a UI component) and unit tested (JSUnit integrated not done - should be easy, though) from a static HTML file.