Introduce badge utilities on the client.
This is still really barebones (I'm just learning Backbone) but is a start for basically sending down JSON for badge info from the server, have it parsed and rendered.

The View is bound to the model so any changes to the badge list will update the UI, as promised by Backbone. Nice job, Backbone.