Introduce the concept of "public user badges" in the models.
This is just saved a serialized list of badge names in the user data, following the existing conventions.

Public badges will be dispalyed in the user's display case without related user information, so it will be super fast to just retrieve the list of public badges (we can just read from the cached list of badges and filter).