get rid of endangered mode - never lose proficiency, just drop into review mode
Endangered mode was originally introduced due to issue 80:
http://code.google.com/p/khanacademy/issues/detail?id=80

Previously, if a user gets 2 in a row wrong on a proficient exercise, they lose
their proficency. Now, the user will never be able to lose proficency, but an
exercise will enter a review state as soon as there's a wrong answer on
a proficient exercise. So, this doesn't completely revert issue 80, because
instead of just reducing the review interval as was done before issue 80, we put
the exercise immediately into review state.

There are various reasons for doing this:
- "Endangered mode" is a confusing state for users
- Being able to lose proficiency may discourage users from doing reviews or
  experimenting
- Simplifies the code
- Reduces # of edge cases we have to handle for the new review mode

Future review algorithms are intended to compensate for this by requiring users
to do more review problems if they get problems wrong after proficiency, instead
of losing proficiency.