First cut at avatar list on the server.
The authoritative list of avatars will be cached in memory (just like badges), since it will probably be an identical limited set for all users. Determining which avatars are accessible to the user will be determined at request time.

Right now there are only "points-based" avatars, but theoretically we can extend it to other types.