Update to struggling model experiment.
This is a re-release of the existing experiment, but now with some suggestions to try pre-requisite execises. (Outstanding TODO to update visual style so that notification is not prone to user blindness).

This also changes some of the metrics tracked to get rid of useless ones and track some ones I'm curious about.

More importantly, it fixes a bug where updating the UserExerciseCache doesn't opt the user into the experiment. Seems like there is a different codepath when going through UserExerciseGraph.get_and_update, versus UserExerciseGraph.generate(). As a result, if a student did a question and were not already opted into an experiment, it would pull from the experiment bucket and not opt them into any experiment. This means any bingos would 404 until they were properly opted in.