Initial commit of NicknameIndex
This introduces entities created for each user that will make it easy to match against search queries.

The basic technique is to tokenize out the names and sort them so that prefix based scans can be done quickly, so that we can also eventually do autocomplete/fuzzy matches when a partial query like 'Salman Kh' is entered (though this is not yet done in this change).

No ranking is done.
No stop words are used (e.g. may want to do "Mr." or "Mrs." and all internationalized forms)