Move some things out of global scope.
As profiles becomes complex, we should move more of the init data into an anonymous JSON passed into the init function. This avoids polluting global scope.

There was already an issue with a name conflict with USERNAME being set on the base page template (for Facebook) and USERNAME on the profile page.