Introduce a UserProfile object.
This is really just rejiggering some of the hard work Marcia did in sending the data down as a UserCardModel. Specifically, I moved some stuff out that didn't belong to any particular profile (e.g. total number of videos in the library), and made a dedicated object on the server.This will make it easier to re-use for viewing multiple profiles or other people's profiles...I think.