Just make profile page not editable for phantom users.
We show a nice banner at the top prompting them to login. We should maybe also pull in some of things we introduce for the out of thebox experience explaining what things are on the page.

But in any case, it doesn't make sense to do things like "mark a profile as public" or "grab a username" as a phantom user, so just deny them for now.