Introduce a "profile-intro" package.
This drops in the code of the guiders v1.2.0 library (sans images, which I don't think we'll need). It is an Apache Licensed project from optimizely that we'll form as the basis of the out-of-the-box experience for profiles: https://github.com/jeff-optimizely/Guiders-JS

Upcoming changesets will involve tweaking that API and actually building out the flow.