Force a transaction apply() after claiming a username.
Since HRD writes happens in two phases (commit/apply), and the second phase is async, it's possible that subsequent queries fetch inconsistent results. For most things on UserData, this has not been a big deal. For usernames, this is a bigger deal since it could result in forever-lost usernames. As a result, force the transaction to flush