Update router, address bar, links upon username change
Known bug: clicking on the points in the upper right corner of the page will still send you to the wrong place because it is a span and not a link. I suspect very few people will click there and get confused, though. I am probably typing more to explain this bug than fixing it. Sigh!