Fix server-side templates for badges.
The GroupedUserBadge object was simplified so that it had a single, flat list of context names, since the concept of what's "visible" and what's "hidden" behind a more link is a UI-specific notion. That logic was unfortunately removed and still required on the server side templates, so I've restored it now in templatetags.