Write video transcripts into page for SEO
Currently doing this simply. In the future this should be integrated with
desmond@'s work on interactive subtitles.

Also, pre-building and caching the HTML in the subtitles update job would
remove the need to build HTML on each page view. However, some quick testing
shows that building the HTML is pretty cheap anyway.