Changes to how profile root/URL routing works.
There was a significant problem before where an empty value in Backbone.History.root caused lots of problems, and didn't actually re-route things properly in IE.

This changes it so that Backbone.History.root is "/profile", which leads to the nice experience that in IE URL's will be "/profile#/username/achievements" and non-IE will be "/profile/username/achievements".

Tested: IE8, IE9, Chrome.