Several bug fixes in PyBars, and improved handling of partials.
Hackaround for PyBars not recognizing {{else}}.
Support for .length paths for arrays.
Automatic partial registration in Python.
Changed the registered partial names to be in a "package_template" format so we can deduce the original filename.
Adding a bunch of template unit tests (.json files)

More PyBars fixes, and a lot of work on partials.