Don't throw if an exercise from an old goal is missing.
This is a stopgap fix that only ensures we don't throw an exception. It results in an odd behavior since the Goal objective is still there without a proper URL (so if you click on it, it will take you to the homepage). Not sure what the right fix is - probably to remove the objective from the Goal, but that's trickier.