Fix an issue where users weren't getting a checkmark for a completed video.
Tasks that update UserVideoCss are not guaranteed to run in order, and I assume when things went bad and tasks were failing and had to retry often, that things did indeed run out of order. This change makes it so that it's acceptable for those tasks to run out of order, and not have the user lose completed status of a video.