Avoid incorrect relative import
In Python2.5 this was ok because the import would start looking at the app
toplevel when the relative module was missing. In Python2.7 an ImportError is
raised when the relative module is not found.

Also, move the import to the file toplevel to avoid import into the global
namespace on each execution of the function.