Re-do the way password change is done.
This introduces an iframe in which the password change form is made in, so that it can simply POST to the https domain, and re-render on errors. If it's part of the profile page, wiring up the messaging about errors or when it succeeds requires wiring that information all the way through the profile handler and all the profile scaffolding in the client, which is a lot messier.

--HG--
rename : javascript/profile-package/settings.js => javascript/login-package/settings.js