Add pep8 lint checking
The blacklist contains all files that don't currently pass pep8, so it can be
run immediately to find problems in new files.  A "lint" target has been added
to the Rakefile as a shortcut. The new "tools" directory is intended for tools
that aren't pushed to GAE to avoid clutter in the app root.