Make /completesignup page robust to having conflicting data.
There were issues where if an existing user is signed into the browser, and the user clicks on a verification link in her e-mail, the existence of both a token and an existing signed in user could give conflicting info. Just to be safe, this makes it so that the token takes precedence, and will not over-write the user data in the existing signed in user.