Move from admin_only to manual_access_checking for urls under /admin.
This is a more conservative version of the access-required changes I'm
making.  Some urls are admin-only by dint of being under /admin, which
is controlled via app.yaml.  app.yaml does a different check than we
do in our own admin_only checking.

I'm not sure if there's a difference or not, but in my testing I saw
some task-queues failing when I had the @admin_only decorator, and
succeed when I changed it to @manual_access_checking.  We can figure
this out later, if/when we decide to move off app.yaml for
permissions.  For now, I'm happy to make this change not-contentful.

The new permissions are manual_access_checking to make it easier to
audit them later, if we *do* decide to move off app.yaml for
admin-testing.