Change the admin-only logic to be the same for API and non-API urls.
The basic difference is the API calls require a valid current-user
object from Khan's point of view (that is, a user in the khan
user-db).  This was used to enforce XSRF cookies, which only matter
for /api/ calls (well, they matter for all POST calls, but we only
check them for /api/ calls currently.)

Checking for a valid current-user even outside of the /api/ directory
tree should be harmless -- in fact, it should be mostly a noop, since
XSRF checking is currently disabled outside /api/.  But unifying the
two APIs is beneficial.

The only time the new logic might give a problem is for task queues,
cronjobs, mapreduces, and the like, which don't run as a khan-academy
user.  Those will have to be protected by app.yaml.  I checked, and no
such operations are currently admin_only, so we should be fine.

Testing: manually checked the (few) admin_only-protected urls on my
dev_appserver instance.  They all worked as expected.