Large change to rip apart /login and /completesignup (the two pages with password forms) into an outer shell and an inner iframe.
The main form contents are now rendered in an inner iframe with the same URL and a &form=1 parameter. This iframe can be rendered in a different domain (with https), and submitted as normal. No cross frame communication must be done this way, except just taking the top frame to /postlogin, which can be done via window.top.location (since that's writeable cross-domain)