Change anointed_oauth_consumer_only to always succeed for cookie-based auth.
It used to be that for an @anointed_oauth_consumer_only handler, if
the user came in via cookies, the test would always fail.  Now it
always succeeds.  So @anointed_oauth_consumer_only is similar to the
old @oauth_optional(anointed_only=True), rather than the old
@oauth_required(anointed_only=True).  That's ok, since we've gotten
rid of oauth_required-ness -- we always allow either oauth or cookies
everywhere.

Reviewed by benkomalo