Don't give an error on @anointed_oauth_consumer_only if not logged in.
The function basically has no logic at all unless it's an oauth
request (code is a bit more twisty than that because I'm not exactly
sure how caching works here).  If it's an oauth request, we look at
flask.g.is_anointed, but if not we'll return True, without even trying
to call verify_and_cache_oauth_or_cookie, which is unhappy for
not-logged-in users.