Fix a bug where duplicated ID's appeared on the homepage and confused content rendering.
This fixes the page for IE8-in-IE7 compat mode, which some people are still using, and is just a generally needed fix since colliding ID's are not good.

The container ID had to be changed since using hashlinks like "#polynomials" was relying on an <a> tag with matching name and id. What's weird is that the data-role="page" with a matching ID overrules an <a> tag with a name attribute only. And for some reason <a> tag with name and matching id overrides the data-role page.