Change the third-party oauth code to create a default request-token object.
We (Khan) have decided not to have different access control for
different oauth requests, so we short-circuited the scope-lookup and
just always use the default request resource.  We might as well go
further and actually *make* the appropriate entry in the db
automatically, so all the clients don't need to make the same thing
themselves.

(Note: the above is based on my understanding of what's going on, and
may be off in details.)