Don't try to use transactions when accessed via the remote API.
remote_api_shell doesn't support queries inside transactions, so we
wouldn't be able to call set_password() from remote_api_shell without
this change.  Note this means that if the request fails, the
remote_api_shell user is going to have to try to roll back (or
correct) the partial state manually.