Reduce some common log-spam: we don't need to know when XSRF isn't being used.
We (arguably) want to log when XSRF actually finds an invalid attempt,
since it may indicate an attack.  But when a user isn't using XSRF at
all: there's neither an XSRF header nor an XSRF cookie value, there's
nothing interesting to log.

This also happens all the time, since we don't require XSRF on many
requests, especially those that are OAuth-only.  I think what's
happening is that if you make an oauth request but aren't in the oauth
map yet (so we start the login process), we'll check whether you have
cookie support, and require XSRF, but you won't have had XSRF tokens
yet, so we'll log a message about a harmless situation.

I could be wrong in my analysis.