Have authorize_token always return a message on error, rather than raising.
I thik the intent of this function is it should return
oauth_error_response(something) if it runs into problems.  The way the
code was, it was doing this for most types of problems, but not all:
for some it would raise an OAuthError that wasn't being caught.  Now I
put the entire function in a try/except block, so we're guaranteed to
catch any OAuthError and return an appropriate error code to the user.