Move code to use ensure_in_transaction, and make that fn remote-api aware.
It was an oversight that the auth/models code wasn't using
ensure_in_transaction in the first place.  By moving over to that, and
also moving the check about whether we're a remote-api client or not,
I make sure *all* code will be able to run via remote-api, since it
all (should) use ensure_in_transaction when appropriate.