Add a test utility to make it easy to fetch KA urls via oauth.
This is intended to be used by unittests.  These tests can use the
utility function here to fetch a url as if we were visiting the khan
website -- including going through all the oauth code-paths in the
appengine app -- but against a dev-appserver.

The first client of these utilities will be end-to-end tests of all
the handlers in /api.

Co-written with ben komalo.