Add support for tests that require a dev_appserver instance to be running.
These are helper routines that start up the dev_appserver instance and
tear it down again.  This is intended to be used for end-to-end tests,
that want to send a url to a server and make sure the results are
appropriate.

As a proof of concept of this new functionality, used it to write a
simple end-to-end test for /api/v1/user.  (This also uses the new
oauth-testing framework.  It all works!)