Get rid of all circular imports within the former models.py files.
For instance, ExerciseVideo uses Exercise, which uses ExerciseVideo.
I resolve this by the use of inside-method imports, which is a big
no-no, but what can you do?  It will be a lot more work to refactor
this so we can remove the circular imports, and may not even be
possible.  This allows us at least to make forward progress.