Fix some more circular import issues and dead imports.
As a fallout of b3076e48ce78 (which was to fix a different circular import issue), templatetags.py was imported at a later stage. It turns out this file is a huge cause of a lot of the circular imports we're having and importing it once out of the way in main.py seems to make a lot of that happier.