Remove topic traces from q&a code
Historical fun fact: there was a day when non-js users were able to view all discussion content through the magic of links-that-worked. That was a great day. Considering that the video doesn't play without JS and this behavior has been broken for a while without complaints, this code is going away.