Fix topic page rendering on prod.
Not sure when/how this broke, but it seems to use camel casing now on prod and the handlebars templates were never updated.

What's a little concerning is that there is no way to invalidate this as we roll out new versions, since this is cacheable. Not sure if that will be an issue later on? Put in a TODO for you, Tom, if that's OK.