Agument transaction_util to allow args/kwargs for the transaction-fn.
Previously, ensure_in_transaction only took functions with no
arguments, making it less powerful than db.run_in_transaction.  Now it
should be equivalent, with the caveat that if the passed-in function
takes an 'xg_on' parameter, it won't be passed on.

I use this to convert a few more uses of db.run_in_transaction to
transaction_util.ensure_in_transaction.