Refactor a bit in order to provide fix_sys_path, a useful utility function.
This function sets the pythonpath (sys.path) to be able to import both
appengine modules and Khan Academy modules.  It also sets environment
variables that appengine needs to function properly.

Any script we run from the commandline that wants to import Khan
Academy code will need to run this.

Probably this functionality should move out of runtests.py, but I
haven't figured out a good way to bootstrap it yet.