A script to make a test db that can be used with dev_appserver to test KA.
This creates all the database entries needed for a Khan Academy
instance.  One can run a test dev_appserver against it and run queries
against it, for end-to-end tests.

This version creates some non-user-related data (videos, exercises,
etc), and a few users.  There are many TODOs, such as all the
user-related stuff.  You can't currently watch videos, and the topic
map seems to be messed up, but the homepage loads!  Or at least it
does if you patch a bug in library.add_next_topic...