Add support for pruning directories to runpep8.
This was messy to do in shell, so I rewrote runpep.sh in python.  (It
moved from being a 27-line shell script, mostly comments, to a 78-line
python script, but ah well.  I think it's easier to follow now, in any
case.)  The new code imports pep8 directly, rather than shelling out
to it, so may actually be a bit less overhead to run.

As proof of concept, ran runpep8.py and cleaned up some of the errors
it found in my new test-db file.