Make each topic-tree layer consistent, and add early maplayout logic.
The actually KA website has the property that every child of a
topic-node is either a list of sub-topics, or a list of
exercises/videos.  You don't have a case where you have some videos
and some sub-topics as a child.  This commit changes the test
topic-trees to have that property as well.

I also added the beginnings of some knowledge-map code so we could
load exercises from this db.  It doesn't work yet, though.