Make profile URL resolution a little smarter.
There are lots of crashes in the log with UTF-8 decode errors. It turns out that the class_profile page is sending users to the individual student profile pages without encoding the profile URL properly. Also, it turns out there are weird e-mails with percent signs in them. When the profile page tries to resolve this, it URI-decodes the content, and the %-sign decodes it into some nonsensical value and we 500 trying to decode it.