Change the behavior for users logging in with Google/Facebook.
This is a subtle, but important change that retrieves users by "user_input_email" if no existing email with the user_id is found. That means that Facebook users logging in with a matching e-mail address to an existing Google account will log into that existing account instead of creating a new account.

This also stops automatically updating Google user emails (with a TODO in the future to allow users to explicitly change that information on Khan Academy)

This is desirable for several reasons:
- it paves a path towards having Facebook accounts have emails!
- it ensures that accounts created on Khan Academy have unique "user input emails"
- it paves a path towards having a single KA account with multiple third party auth providers