Set up pythonpath correctly for third-party files.
Here's some necessary background, using agar/ as an example:

We don't want to just have people import from the agar source directly
in third_party, because due to the way the agar directory tree is set
up, to import from it you'd need to do
   import third_party.agar.agar
rather than just
   import third_party.agar
(this is because we don't install agar via 'pip install' or the like,
but instead just use the source directly).

Even without that problem, agar code internally does
   import agar.foo.bar
which would fail normally, since we put agar under third_party.agar,
not agar.

Instead we install the source under agar-src (using a - instead of a _
so people can't import from it accidentally), and have a symlink from
agar to agar-src/agar.  That allows
   import third_party.agar.foo.bar
and all is happy.

Except the agar package itself isn't happy, because code in there has
   import agar.foo.bar
and that needs to work.  This file fixes that up by adjusting the
sys.path so it *does* work.  It also adds a tricky import hook to
ensure that nobody makes use of this new sys.path functionality
except for code in third_party -- 'normal' Khan Academy files
need to do import third_party.agar.boo.bar instead.