Update support for the test db, and start using it for testing.
Made a test-db and actually checked in it, so folks can start using
it.  If we don't want a file this big to be in source control, it
compresses really well (and really quickly) -- that support would be
easy to add.

Changed the handler-test-utils to be able to use the test-db rather
than hard-coding in the use of an empty db.

Added a few test-cases that use the new database.

Also added a TODO at benkomalo's request, based on previous review.