Every place we call TopicVersion.get_by_id, return 400 if we get back None.
Before this change, we would not error-check the return value of
TopicVesion.get_by_id.  This means that we would pass in version=None
to subsequent API calls in the given function.  Sometimes this would
result in an empty response, but more often version=None would be
treated as "the current version" (probably not a great design
choice).  So you would silently get back info about the current
version if you entered in an invalid version-id.  Now you get a 400
error, which is better.