Added @login_required_and(), and rewrote existing decorators in terms of it.
This is the most flexible of the authorization decorators (api-only,
for now).  It makes it easy -- well, easier -- to have more complex
authorization requirements.  It also is where we can put simple but
relatively rare auth requirements (like demo_user).

The reorg means we no longer need to use the user_util routines for
figuring out auth -- it's simpler to just look at the user_data object
directly.  As a result, we no longer get the advantage of caching the
auth settings, which was probably not a win anyway since
UserData.current() is already cached, and that's the only part of
auth-checking that would be expensive.