Fix etag-caching for get_topic_page_data.
It's confusing, but here's my understanding:

We don't want to set the etag if the topic-version is the current
edit-version, since that can change without warning.  Unfortunately,
it's slow to check if version_id is the current edit-version or not.
Instead we do the fast check of whether version_id is None (None ==
active version which is never the edit-version).  We disable etag
caching in all other cases by having the etag functor return None.