Start moving toward a world where we set xsrf cookies by default.
When the khan academy app makes API calls in order to fulfill a url
request, it needs an xsrf cookie for authentication.  In the past, we
used to set xsrf cookies only when needed -- that is, when visiting a
url that needs to make API calls behind the scenes.

But that's too complicated to get right, and it doesn't save us very
much: at most, it saves us from setting a cookie in the case a user
doesn't visit any url that makes an API call.  So we go to a new
scheme, which is we set the xsrf cookie any time we visit a page that
requires a login (which includes being a phantom user).

This still means we don't auto-set a cookie if a user visits only the
no-login-required parts of the site -- viewing the homepage, for
instance -- but we do otherwise, and most of the time.

Note that this is just scaffolding: no handlers use @login_required
yet, so this functionality is unused.

Testing: make allcheck, and poking around the site a bit manually.n