Another place we need to force-commit to get the HRD to see a commit.
When running a dev-appserver, I was consistently (100% of the time)
seeing 'failed to create phantom user' when filling out an exercise
while not logged in.  After much debugging, I tracked it down to this
put() not taking place right away.

We believe that this happens in production as well -- albeit rarely.
Ben Komalo has stated that this shouldn't be too expensive to add,
since it is only called when creating a new user, which happens at
most once per session.