@login_required is not the right decorator to use with @api_create_phantom.
Here's why (it's complex and annoying): @api_create_phantom wants to
'log in' the user as a phantom user, but only if they are not
currently logged in.  But in API-land, we don't know if the user is
logged in until @login_required or one of its friends is called.  That
means @login_required has to run before @api_create_phantom runs.  But
if @login_required runs first and the login fails, then login_required
throws an exception, causing the request to fail.  That's bad because
what we *want* to have happen is to create a phantom user in that
case.

The "right" fix is probably to make login_required take a
'create_phantom_if_required' flag, and thread that through the logic.

However, there's another fix which also works, and is much easier: use
@open_access, which works exactly like @login_required in that it
tells the API system who the logged-in user is, but is unlike
@login_required in that if there is no logged-in user, it just
continues silently rather than failing.  With that, we can do
@open_access first, and then @api_create_phantom second, and the
@api_create_phantom will log us in as a phantom user if it turns out
the @open_access log-in failed.