Consolidate getting user data from parameters some more and have it support username parameters.
This somewhat mimicks some of the functionality in UserData.get_possibly_current_user, but is slightly different. In the future, I'd like to consolidate those methods in UserData as well, but don't want to deal with it for now. The problem with those methods is that if an identifier is specified and it doesn't match the current user, it does a series of DB hits looking by user_id, then username/email. It should really just done one hit based on the name of the parameter, since the client is usually explicit about what type of identifier it uses.

That being said, it probably makes more sense in the future to consolidate this retrieving-of-user-by-explicit-identifier up in a higher level application layer instead of the model object itself.

Also add in a unit test.