Don't keep going if we fail to create a phantom user.
We're still trying to figure out why this might fail.  Our best theory
now is that xsrf is failing, possibly due to the fact that these
requests are coming in by script kiddies.  To test this, add some
logging when we fail to create a phantom, saying what the headers
are.  At least that will let us test the 'xsrf is responsible' theory,
even if it doesn't tell us why xsrf is failing here.