Move gdata to third_party.
This is part of the plan to move all code whose 'main' home is outside
this repository, into third_party.  When possible, we will bring in
this code via sub-repos (by way of a khan-academy fork that we make),
allowing for easier tracking of updates, especially by letting the
source control system keep track of local vs remote changes.

In the case of gdata, it looks like we're using an old version that
sal imported a long time ago.  This updates to the latest stable
release.

There have been only three changes that have been made to gdata since
it was installed: b3be77a6850f, c0982264e938, and 15ccdb119072.  I
will re-make them in a subsequent commit.