Don't do so much logging on can't-get-login-credentials errors.
It's not uncommon for us not to be able to find login credentials,
because someone is accessing an API without having logged in first.
(Either by logging in to KA and getting a cookie that says they're
logged in, or by getting a valid oauth token, which requires having
logged in.)  Those don't need to be logged, since there's nothing we
can do about them.

At least, that's the case if the code is working perfectly.  However,
a zillion of these errors indicates the problem is with our
authentication system, not with the users doing something weird.  So
as I compromise, I still log these failures, but as warnings rather
than errors.  I've added a TODO to have a better notification system
here instead.