Add a wrapper around pickling that can handle classes moving around.
We've had several production problems caused by classes getting
renamed (possibly just by moving directories, since the module-path is
part of a class-name).  The problem is when an object was pickled with
the class in its old location, and then we try to unpickle it after
the object has moved to its new location.  The pickling code sees
OldLocationClass and tries to instantiate it and gets an error, since
OldLocationClass doesn't exist anymore.  Chaos ensues.  (We can't even
roll back the deploy, since then we have a problem with all the people
who created new db entries with NewClassLocation, who we can't
unpickle now that we're back to OldClassLocation!)

(The problem was made even worse by bad code that we took from someone
else which would cause data to be doubly-pickled when we had errors
unpickling, causing corruption we couldn't fix without manual database
tampering.)

Luckily for us, python provides a mechanism to change what the
unpickler does when it sees a class name.  We hook into that mechanism
to make it so we can handle pickled data under either the old *or* new
location.  That should keep this problem from recurring ever again.

Since I had a wrapper anyway, I added some other goodies, like always
using cPickle, and always using a fast/efficient pickling protocol.

I rewrote all our code -- but not third-party code -- to use the
pickle-wrapper rather than the pickle class directly.

I got rid of a hack we had put in previously to work around a
production error with classes moving, which should be obsoleted by
this new class.

Tested: make check (including a new test I wrote)