No longer silently accept unpickle-able data in ObjectProperty.
The code used to read 'if we can't unpickle something, just store it
directly -- maybe it will still be useful.'  But it didn't mark that
the object was corrupted, and this led to objects being pickled
twice.  Here's how:

1) We stored an object (correctly pickled) as an ObjectProperty in the
datastore.

2) We rewrote the code in such a way that the object couldn't be
unpickled (deleted a class referenced in the object, say).

3) The code was like, "Oh well", and just returned the object as a
string.

4) We went to re-store the object, and the ObjectProperty code pickled
it (the string) and stored it.  Now the object is double-pickled.
Even if we reverted the code that caused the failure in step (2), we
can not get at this object again without manually munging the
datastore.

The problem here was in step (3) -- if we can't unpickle something
that we pickled in order to store, it's an error, not something to
silently ignore.  The new code does that properly.

While in the area, I changed the two versions of object_property (one
in gandalf) to match each other, and fixed a typo or two.